Split the reflection and light sheen calculations
The previous approach assumed that sheen and reflections were always both enabled at the same time, since the addReflectionSheen function relies on calcReflections.
Add the useSheen boolean to the public API of the material. This will toggle the sheen effect on and off.
