+++ src/graphics/program-lib/programs/lit-shader.js
@@ -838,8 +838,12 @@ class LitShader {
                 code += chunks.refractionPS;
             }
             if (options.sheen) {
-                code += chunks.sheenCharliePS;
+                code += chunks.reflectionSheenPS;
             }
+        }
+
+        if (options.sheen) {
+            code += chunks.lightSheenPS;
         }
 
         // clustered lighting
